requirements.txt
cachelib==0.1.1
click==7.1.2
cs50==6.0.2
docopt==0.6.2
docx-mailmerge==0.5.0
Flask==1.1.2
Flask-Session==0.3.2
future==0.18.2
itsdangerous==1.1.0
Jinja2==2.11.2
lxml==4.6.2
Markdown==3.3.3
MarkupSafe==1.1.1
num2words==0.5.10
numpy==1.19.5
pandas==1.2.0
Pillow==8.1.0
python-dateutil==2.8.1
pytz==2020.5
sh==1.14.1
six==1.15.0
SQLAlchemy==1.3.22
sqlparse==0.4.1
termcolor==1.1.0
tqdm==4.55.1
Werkzeug==1.0.1
xlrd==2.0.1
